# perf: use fast api for `core.isProxy`