--- a/bench_util/benches/op_baseline.rs
+++ b/bench_util/benches/op_baseline.rs
@@ -41,16 +41,6 @@ fn bench_op_async(b: &mut Bencher) {
   bench_js_async(b, r#"Deno.core.opAsync("op_pi_async");"#, setup);
 }
 
-fn bench_is_proxy(b: &mut Bencher) {
-  bench_js_sync(b, r#"Deno.core.isProxy(42);"#, setup);
-}
-
-benchmark_group!(
-  benches,
-  bench_op_pi_json,
-  bench_op_nop,
-  bench_op_async,
-  bench_is_proxy
-);
+benchmark_group!(benches, bench_op_pi_json, bench_op_nop, bench_op_async,);
 
 bench_or_profile!(benches);