--- a/core/ops_builtin.rs
+++ b/core/ops_builtin.rs
@@ -191,7 +191,7 @@ fn op_format_file_name(file_name: String) -> String {
   format_file_name(&file_name)
 }
 
-#[op]
+#[op(fast)]
 fn op_is_proxy(value: serde_v8::Value) -> bool {
   value.v8_value.is_proxy()
 }