--- a/core/runtime.rs
+++ b/core/runtime.rs
@@ -3121,6 +3121,7 @@ assertEquals(1, notify_return_value);
     runtime.execute_script("<none>", "").unwrap();
   }
 
+  #[ignore] // TODO(@littledivy): Fast API ops when snapshot is not loaded.
   #[test]
   fn test_is_proxy() {
     let mut runtime = JsRuntime::new(RuntimeOptions::default());