Extracted from #15521 

```
main:
  is_proxy               29.26 ns/iter   (27.21 ns ... 33.88 ns)  30.83 ns  31.19 ns  31.32 ns
  response_new_headers  911.85 ns/iter (906.53 ns ... 944.67 ns) 912.07 ns 944.67 ns 944.67 ns

this patch:
  is_proxy               10.81 ns/iter    (9.57 ns ... 15.75 ns)  12.78 ns  13.22 ns  13.28 ns
  response_new_headers  881.89 ns/iter (874.26 ns ... 914.93 ns) 884.52 ns 914.93 ns 914.93 ns
```